Support local S3 compat and move config check
